pyproject.toml
[project]
name = "vuln-mgmt"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "pydantic",
    "pydantic[email]",
    "celery",
    "redis",
    "python-jose[cryptography]",
    "bcrypt<4.1.0",
    "passlib[bcrypt]",
    "python-multipart",
    "structlog",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
]

# ADD THIS BLOCK BELOW:
[tool.setuptools.packages.find]
include = ["app*"]